moving metadata controller lower down
